Character controller reworked
Reworked the character controller due to some major issues regarding
collisions.
* The character controller now casts rays from all his perimeter. The
direction of all these rays is the movement of the frame.
* Removed the properties of CharacterControllerState. It now only stores
if there has been collisions in this frame, if the entity is grounded
and in which platform it's grounded.
* Now the GameControllerInput sets the right properties on the CCCP. The
CCCP now uses these values.
* Replaced the property CanJump with a method with the same
functionality.
* Now the number of rays casted and the skin width are both public
attributes modifiables from the inspector.
* Removed all use of global axis while calculating the physics on the
CCCP. Now it takes as reference the gravity instead of vector constants
(such as Vector3.down). The input still uses the global axis.
* Now the method Jump checks if it can jump instead of hoping that it
has been called before.
* Removed all methods handling slopes. Now slopes are easily handled by
the same method which checks the collisions.
* Now the clamping of the speed is performed on each axis and before the
actual movement.
* Succesfully added sliding on slopes.