Input reworked
Reworked how the character controller uses the input.
* Now the GameControllerInput will also get the vertical axis. The
CharacterControllerCustomPlayer will use it as well.
* Now the parameters for the entity acceleration and speed are in
CharacterControllerParameters rather than the
CharacterControllerCustomPlayer. The CharacterControllerCustom will use
this parameters now.
* Added the parameter MovementFreedom, which lets the player move in any
or both axis.
* Added the parameter MovementBehaviour which lets the player move only
in specific situations. This prevents the player from moving while
sliding.
* Added the RelativeToGravity option which makes the player input be
relative to the gravity rather than the gloval axis.
* Now the facing direction in CCCP is a Vector3 rather than a boolean.
* Added several methods to the CharacterControllerCustom regarding the
movement relative to the gravity.
* Added the method SetInputForce to the CharacterControllerCustom which
calculates the right velocity of the entity depending on it's
parameters.
* Now the jump direction is always relative to the gravity.
* Fixed a problem which allowed the entity to be grounded and sliding at
the same time.
* Now OnCustomCollision is called before any other collision method.