Slide on vertical walls
Improved the movement of the character when sliding along slopes. The
character will stop when colliding with a slope and then it will fall
slower than before.

* Added a new property to CharacterControllerState: IsFalling. If true,
the character was falling on the previous frame.
* Added a new jump behaviour: CanJumpSliding, allowing the character to
jump while sliding. This is now the default value.
* Added thre properties to CharacterControllerParameters:
- maxWallSlideAngle, that defines the maximum angle of a wall the
character can slide among.
- slidingDragFactor, that defines the reduction on gravity when the
character is sliding down a slope.
- angleThereshold, a value used when comparing float angles. Used for
stability.
* The CantMoveSliding movement behaviour now allows movement if the wall
angle is close to the maximum slide angle. Otherwise, it won't be
possible to slide down vertical walls.
* Now the character will only be considered sliding when the slope angle
is beetween the slope limit and the maximum wall angle.
* Now the character will stop when first colliding with a slope.