Wall jump
Added wall jumping to the character controller. After wall jumping, the
character will be unable to move for a short time period.

* Added a new property to CharacterControllerState: IsOnSlope, which
will be true if the character is standing on a slope, sliding or not.
* Added a new movement behaviour: CantMoveOnSlope wich will prevent the
character from moving while standing on a slope, sliding or not.
* Added a new jumpbehaviour: CanJumpOnSlope wich will allow the
character to jump while standing on a slope, sliding or not.
* Added the next properties to CharacterControllerParameters:
- wallJumpMagnitude, the jump height reached by a wall jump.
- wallJumpFlyTime, amount of time the character will be unable to move
after performing a wall jump.
* Now the character acceleration won't be modified by the character's
size, making it more tricky to control bigger characters.
* Now the sliding drag factor will be reduced by the character's size,
making bigger characters slide faster down slopes.
* Now the player will be unable to jump before releasing the jump button
beetween jumps.
* Made the next changes to the prefab configuration:
- Decreased the air acceleration to 3.