Sphere deform
Created a new deform script based on a sphere. The script has the same
properties as the box one, but allows deformation from different
direction, even simultaneously.

The next changes has also been made:
* Removed the mesh from the character game object. Created a new
children for the character that stores the mesh.
* Now the mesh object is slighty bigger than the character collider in
order for the deformation to be applied.
* Added a trigger collider to the mesh's game object as well as the
deform script.
* Changed the LateUpdate method for Update on CharacterControllerCustom.
The deformation needs to be performed after the actual movement so it
will be called on LateUpdate.
* Fixed some spelling errors on comments and summaries on
CharacterControllerCustom.