Alternative Smarter Sphere Deform
Created an alternative sphere deform script. This one will fit better
the space and will not overlap with the world. However, this version
will NOT keep the volume of the sphere, meaning the deformation doesn't
look as good as before.

* Changed the number of caster days from 32 to 16 on the character
prefab. The entities on the scene still use 32 for testing purpouses.
* Changed the Awake function to Start on the deformation scripts.
* Added OnEnable to the deformation scripts to restart the deformation
when the entity is re-enabled.